@@ -10,17 +10,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.0.0",
-    "@angular/common": "^14.0.0",
-    "@angular/compiler": "^14.0.0",
-    "@angular/core": "^14.0.0",
-    "@angular/elements": "^14.1.0",
-    "@angular/forms": "^14.0.0",
-    "@angular/platform-browser": "^14.0.0",
-    "@angular/platform-browser-dynamic": "^14.0.0",
-    "@angular/router": "^14.0.0",
+    "@angular/animations": "^14.1.1",
+    "@angular/cdk": "^14.1.1",
+    "@angular/common": "^14.1.1",
+    "@angular/compiler": "^14.1.1",
+    "@angular/core": "^14.1.1",
+    "@angular/elements": "^14.1.1",
+    "@angular/forms": "^14.1.1",
+    "@angular/material": "^14.1.1",
+    "@angular/platform-browser": "^14.1.1",
+    "@angular/platform-browser-dynamic": "^14.1.1",
+    "@angular/router": "^14.1.1",
     "@ngrx/store": "^14.0.2",
-    "generic-list-filter": "^0.1.6",
     "ng2-search-filter": "^0.5.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
@@ -29,7 +30,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.1.1",
     "@angular/cli": "~14.1.1",
-    "@angular/compiler-cli": "^14.0.0",
+    "@angular/compiler-cli": "^14.1.1",
     "@types/jasmine": "~4.0.0",
     "jasmine-core": "~4.1.0",
     "karma": "~6.3.0",
